Rename 'customExceptions' package 'exceptions'.
