+++ src/com/lms/exceptions/DeleteException.java
@@ -1,21 +1,21 @@
-package com.lms.customExceptions;
+package com.lms.exceptions;
 
 /**
- * An exception class to report the failure of an 'insert' operation from the
+ * An exception class to report the failure of a 'delete' operation from the
  * service layer to the application layer.
  *
  * @author Salem Ozaki
  * @author Jonathan Lovelace
  */
 @SuppressWarnings("serial")
-public class InsertException extends TransactionException {
+public class DeleteException extends TransactionException {
 	/**
 	 * To throw an instance of this exception class, the caller must supply the
 	 * exception message.
 	 *
 	 * @param errorMessage the exception message
 	 */
-	public InsertException(final String errorMessage) {
+	public DeleteException(final String errorMessage) {
 		super(errorMessage);
 	}
 }
