+++ src/com/lms/exceptions/InsertException.java
@@ -1,21 +1,21 @@
-package com.lms.customExceptions;
+package com.lms.exceptions;
 
 /**
- * An exception class to report the failure of an 'update' operation from the
+ * An exception class to report the failure of an 'insert' operation from the
  * service layer to the application layer.
  *
  * @author Salem Ozaki
  * @author Jonathan Lovelace
  */
 @SuppressWarnings("serial")
-public class UpdateException extends TransactionException {
+public class InsertException extends TransactionException {
 	/**
 	 * To throw an instance of this exception class, the caller must supply the
 	 * exception message.
 	 *
 	 * @param errorMessage the exception message
 	 */
-	public UpdateException(final String errorMessage) {
+	public InsertException(final String errorMessage) {
 		super(errorMessage);
 	}
 }
