+++ src/com/lms/menu/BorrowerMenu.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.Optional;
 import java.util.OptionalInt;
 
-import com.lms.customExceptions.TransactionException;
+import com.lms.exceptions.TransactionException;
 import com.lms.model.Author;
 import com.lms.model.Book;
 import com.lms.model.Borrower;
