+++ src/com/lms/service/BorrowerServiceImpl.java
@@ -13,10 +13,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
-import com.lms.customExceptions.DeleteException;
-import com.lms.customExceptions.InsertException;
-import com.lms.customExceptions.TransactionException;
-import com.lms.customExceptions.UnknownSQLException;
 import com.lms.dao.BookLoansDao;
 import com.lms.dao.BookLoansDaoImpl;
 import com.lms.dao.BorrowerDao;
@@ -26,6 +22,10 @@ import com.lms.dao.CopiesDaoImpl;
 import com.lms.dao.DBConnectionFactory;
 import com.lms.dao.LibraryBranchDao;
 import com.lms.dao.LibraryBranchDaoImpl;
+import com.lms.exceptions.DeleteException;
+import com.lms.exceptions.InsertException;
+import com.lms.exceptions.TransactionException;
+import com.lms.exceptions.UnknownSQLException;
 import com.lms.model.Book;
 import com.lms.model.Borrower;
 import com.lms.model.Branch;
